log_receive.m:    Added code tracking parameters and I2Q2 values. data_feed.cpp:    Added segment sequence number. DE2_TOP.v, rt_data_feed.v, rtdf_packet_processor.v:    Added missed packet checking via sequence number. channel.v:    Connected code tracking and tested on data with 0 code shift.
Still experiencing strange carrier tracking blips. First presumed from packet loss, but this doesn't appear to be the case. The loss may be because of unforseen data rate mismatches causing the data feeder to stall, however there don't seem to be missed packet events during blipped logs. Maybe there is a bug in the data feed code that causes it to discard a word?